Created a unit production weighting list so CPU player knows which units it should prioritize when producing new units. Some issues, it doesn't know not to make prototype units. There is no randomness/variation so will produce same units most of the time. Modified prefabs so gamePieceId is an int instead of a byte, finally. Realized that GameValuesSO.LoadUnitsFromCSV was replacing the gamePieceId for units to an increment instead of what it read from the .csv, was causing issues with the CPU weighting function.

+++ SO-Client/Assets/Scripts/BaseUnit.cs
@@ -16,7 +16,7 @@ public class BaseUnit : MonoBehaviour
     //enums as defined in AttributesBaseUnit
     public Progeny progeny; 
     public UnitTerrainType unitTerrainType;
-    public byte gamePieceId;
+    public int gamePieceId;
     public bool isNotPrototype;
     public HealthType healthType;
     public DamageType damageType;
